Disallow sharing zstd (de)compressor contexts
According to the zstd author, it is not possible to share Zstandard objects across thread boundaries. To resolve this, we check if the object was created on the current thread and raise a RuntimeError if it is not.

The tests are updated to ensure that the error is raised if a (de)compression context is shared across threads.