Disable new stack size code for non-glibc.
MUSL claims to support pthread_getattr_np, but the response returned is
off by at least an order of magnitude compared to reality.  This is
apparently a long standing problem.  MUSL also does not provide any way
to detect directly that it is the libc.  It would be possible to add
things to configure to detect that MUSL is the compiler, but since this
code has caused problems for a lot of non-glibc platforms, it seems
easier to just restrict it to glibc.