gh-134644: handle exceptions set in `PyOS_Readline`
The builtin input calls `PyOS_Readline` but seems to assume it does not set
exceptions: if the call fails it checks signals and runs handlers if any are
pending, which will cause an assertion failure if an exception has already been
set.

Fix this by only checking signals if an exception has not already been set.