gh-116909: fix data race with versions in typeobject
Global state `_PyRuntime.types.next_version_tag` is being accessed without
synchronization or atomics. This could potentially result in the same version
being used in two different type objects, incrementing past the maximum limit,
and the usual non-atomic memory access issues.

Fix this by using atomics to ensure the version is accessed and updated in a
race-free manner, while also ensuring it is never incremented past the expected
(maximum + 1).

Note there is a theoretical change in behaviour with the second use, for static
builtin types, if their versions exceed the maximum, with assertions disabled:
previously they would have continued incrementing past the maximum and using
the increasing version number; now they will all use the maximum. It might be
better to replace the `assert` with an `abort()`: I assume we would be crashing
shortly if we continue after this, both before and after this change.