Assert no exception is raised in PyErr_CheckSignals
If a signal has been received `PyErr_CheckSignals` will run its handler, if
any. Since that involves calling a Python function it requires there not be an
existing exception raised. However this is not checked until we try and make
the call.

Add an explicit assertion check to this function to ensure the precondition
holds, regardless of whether a signal is pending or not. This will detect any
unsafe calls regardless of signal status.