Revert all commits after "Use threading.excepthook instead of sys.excepthook"
This reverts this branch back to just calling `threading.excepthook`, instead
of trying to raise the exception when the user gets the result or exits the
context manager. I would like to get the immediate bug fixed, and I think this
is a simple, low-risk fix that does that.

From what I understand it isn't possible to prevent the reference cycle which
is causing all those test failures. We need to store the exception to raise it
later, which implicitly creates one.

We can always add raising the exception as an additional improvement later, but
in the meantime if any user code wants to handle the exceptions it can do so by
installing a default exception handler.