gh-117888: print failing command when platform triplet detection errors
The build system determines the platform triplet using the C preprocessor. If
that fails for whatever reason we presently just report "none" for the triplet
and continue until the invalid triplet causes a messy failure down the line.

Instead, store any error output, and in the event of a failure send it to
stderr and fail immediately with a clear error message.