Replace _Py_ALIGN_AS(V) by _Py_ALIGNED_DEF(N, T)
This is a common façade for the various `_Alignas` alternatives,
which behave in interesting ways.

The standard `alignas` errors if it would decrease the alignment;
to prevent that it can be used again with the defined variable's
type.

The standard `alignas` can't be used on a `struct` definition,
the workaround is to use it on one of the `struct`'s members.

MSVC `declspec(aligned)` only takes integer literals.

MSVC `declspec(aligned)` had a bug when applied to a combined
struct+member definition.
The workaround is to separate the struct definition.
Do that for `PyASCIIObject.state`.