gh-127146: Emscripten: Fix pathlib glob_dotdot test
The Emscripten path resolver uses the same mechanism for resolving ..
at a file system root as for resolving symlinks. This is because
roots don't store their mountpoints. If the parent of a node is itself,
it is a root but it might be a mountpoint in some other file system.

Anyways, if a path has enough ..'s at the root, it will return ELOOP.
Enough turns out to be 49.