Make iselement stricter + incorporate feedback
Adjusts iselement() based off feedback from picnixz. Since ElementTree
needs to be compatible with element-like objects, the function now checks
for all attributes that are necessary to instantiate an ElementTree object
and successfully write it to a file (preventing issue #135640).
This commit also incorporates other feedback requesting moving the
tests and adding more tests, and adds a news entry.