Use inspect.isroutine() in DocTest's lineno computation
Previously, DocTest's lineno of `functools.cache()`-decorated functions
was not properly returned (None was returned) because the underlying
computation, in `DocTest._find_lineno()`, relied on
`inspect.isfunction()` which does not consider the decorated result as a
function.

We now use the more generic `inspect.isroutine()`, as elsewhere
in doctest's logic, thus fixing lineno computation for
functools.cache()-decorated functions.