gh-127146: Emscripten: Set umask to zero in python.sh (GH-136740)
Clears the umask used during a test of pydoc.apropos when testing on
Emscripten. This is to work around a known issue in Emscripten; but it's not
clear if the chmod call that is causing the problem is actually testing
anything of significance.
(cherry picked from commit 22af5d35a620ee44393853036a8450ceb047688e)

Co-authored-by: Hood Chatham <roberthoodchatham@gmail.com>