Fix race condition in the test case (for real this time).
Apparently, the thread can start waiting at a point where the server
isn't ready to receive messages, which caused it to block indefinitely
for some reason. Before we create any threads or set any events, do a
simple exchange with the server to ensure that it's in its message loop.
This took way too damn long to figure out.