Python - Windows: Fix Unicode errors from AI generated code from causing an abort
With the rise of AI generate code, python Windows is suffering from this bug at an exponentiate rate. I hope that this commit can be accepted is, or a conversation can be had about the proper target for a fix like this. Perhaps `textio.c` is too coarse grained, and it needs to be an attribute to the print function (and logging).